requirements.txt
Django==5.0.6
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.7
rest_framework_simplejwt==0.0.2
